requirements.txt
python-can >= 4.4.2